+++ src/utility/useAppState.ts
@@ -9,6 +9,7 @@ export type AppState =
   | 'singledevrepostats'
   | 'barbarianrankings'
   | 'mailmap'
+  | 'longevity'
   | 'about'
   | 'exit'